requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==0.3.0
